requirements.txt
discord.py
dotenv
pathlib
psutil
asyncio